+++ components/layouts/AdminFooter.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { Layout } from "antd";
+const { Footer } = Layout;
+
+export default function AdminFooter() {
+  return (
+    <>
+      <Footer style={{ textAlign: "center" }}>
+        Ant Design ©2023 Created by Ant UED
+      </Footer>
+    </>
+  );
+}
