+++ pages/_document.tsx
@@ -3,7 +3,7 @@ import { Html, Head, Main, NextScript } from "next/document";
 export default function Document() {
   return (
     <Html lang="ko">
-      <Head></Head>
+      <Head />
       <body>
         <Main />
         <NextScript />
